Return hdrSize in ListPack

+++ internal/restic/repository.go
@@ -32,7 +32,10 @@ type Repository interface {
 	//
 	// The function fn is called in the same Goroutine List() was called from.
 	List(ctx context.Context, t FileType, fn func(ID, int64) error) error
-	ListPack(context.Context, ID, int64) ([]Blob, int64, error)
+
+	// ListPack returns the list of blobs saved in the pack id and the length of
+	// the the pack header.
+	ListPack(context.Context, ID, int64) ([]Blob, uint32, error)
 
 	Flush(context.Context) error
